Instruments and some suggested fixed
Added Clarinet, Trombone, and Trumpet sounds for  a total of 5
instruments! Wasn't able to figure out how to switch instruments - might
require us to fully re-initialize the BlipGenerator and its cells.

Changed the Load/Save page to prompt the user if they are overwriting a
file or deleting a file.

BUG - Loading is still on/off working