Instrument refactors, swing functionality, sad day :(
Changed some of the scale orderings to better reflect the 7 scale modes
in order (and reference them by name). Redid the trombone and trumpet
sounds so they have more of an attack and are shorter. Labeled Sax notes
correctly. Added swing functionality. Unfortunately, I somehow forgot
the password to the keystore... so we now have to republish in order to
update. shit